requirements.txt
fastapi
mysql-connector-python
uvicorn
pydantic
line-bot-sdk==2.4.2